+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes
 -----------------------
 **--02/19/2026--**
 
+* server:<br />don't resolve master every time gamestat is sent
+
 * server:<br />more constants
 
 **--02/18/2026--**